requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
ultralytics
mediapipe
